pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiprobe"
version = "1.0.0"
description = "AI security testing framework covering RAG, Agent/Tool-Use, and Multi-Modal attack surfaces"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Mongoose Security Team"}
]
dependencies = [
    "openai>=1.12.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "Pillow>=10.2.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
aiprobe = "aiprobe.cli:main"

[tool.setuptools.packages.find]
include = ["aiprobe*"]
